Move Account and ATMProgram into seperate files
