+++ ATMSimulator.sln
@@ -5,11 +5,6 @@ VisualStudioVersion = 17.8.34525.116
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ATMSimulator", "ATMSimulator.csproj", "{BA55A289-6531-4874-90F5-D112087074A1}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{00FE46AD-C6A5-475D-8547-1AC38E525F77}"
-	ProjectSection(SolutionItems) = preProject
-		MyClasses.cs = MyClasses.cs
-	EndProjectSection
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
