+++ Form1.cs
@@ -136,135 +136,4 @@ namespace ATMSimulator
         }
     }
 
-
-
-    public  class ATMProgram
-    {
-       // private Account[] ac = new Account[3];
-
-        List<Account> accounts = new List<Account>();   
-
-
-
-        
-
-
-        public ATMProgram()
-        {
-           // ac[0] = new Account(300, 1111, 111111);
-          //  ac[1] = new Account(750, 2222, 222222);
-          //  ac[2] = new Account(3000, 3333, 333333);
-
-            accounts.Add(new Account(300, 1111, 111111));
-            accounts.Add(new Account(750, 2222, 222222));
-
-            accounts.Add(new Account(3000, 3333, 333333));
-
-
-
-
-
-
-        }
-
-        public List<Account> getAllAccounts()
-        {
-            return accounts;
-        }
-
-        public void addAccount(Account account)
-        {
-            accounts.Add(account);
-        }
-
-
-
-        public Account findAccount(int accountNum)
-        {
-           
-
-            for (int i = 0; i < this.accounts.Count(); i++)
-            {
-                if (accounts[i].getAccountNum() == accountNum)
-                {
-                    return accounts[i];
-                }
-            }
-
-            return null;
-        }
-
-
-    }
-
-    public class Account
-    {
-        private int balance;
-        private int pin;
-        private int accountNum;
-
-        public Account(int balance, int pin, int accountNum)
-        {
-            this.balance = balance;
-            this.pin = pin;
-            this.accountNum = accountNum;
-        }
-
-
-        public int getBalance()
-        {
-            return balance;
-        }
-        public void setBalance(int newBalance)
-        {
-            this.balance = newBalance;
-        }
-
-
-        public Boolean decrementBalance(int amount)
-        {
-            if (this.balance >= amount)
-            {
-             
-               
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-
-        public Boolean checkPin(int pinEntered)
-        {
-            if (pinEntered == pin)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-        public int getAccountNum()
-        {
-            return accountNum;
-        }
-
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
