--- MyClasses.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-
-namespace MyClasses
-{
-    public class MyClasses
-    {
-      
-   
-    }
-}
